+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 [unreleased]
+
+#### 1.3.4 / 2022-03-21
 * run `restore_temp_backup()` in `shutdown` hook
 
 #### 1.3.3 / 2022-03-18